This check-in fixes the following:
In clearCacheSlotSetTo(), set status = 0 when calling uregex_setText to ensure that the value of status is set to a known good value before being passed to uregex_setText.  This probably wouldn't cause any problems in practice, but this makes sure that it can't.

Adds the following:

C Pre-Processor Option Flag 'RKL_REGISTER_FOR_IPHONE_LOWMEM_NOTIFICATIONS'.

This causes some extra code to be compiled in which will dynamically try to register for the 'UIApplicationDidReceiveMemoryWarningNotification' notification.  This is sent in iPhone applications when memory is running low.  The action taken by RegexKitLite is to flush all cached information, reducing its memory footprint as much as possible.  This includes flushing all cached compiled regular expressions and freeing the 'large' dynamically sized UTF16 conversion cache (if any).

The notification registration relies on __attribute__ ((constructor)), which causes the linker to call rkl_RegisterForLowMemoryNotifications() at load time.  rkl_RegisterForLowMemoryNotifications() will only register for the notification once, and once only, as the variable rkl_HaveRegisteredForLowMemoryNotifications is atomically updated with a compare and swap.  Thus, it is safe to call multiple times, from different threads, although the function is delcared static so it is not visible outside the translation unit.

This change also rearranged the string cache clearing code so that it's a bit more modularized.  By default, when the notification is received, the cache spin lock will be 'tried' to be locked.  On success, the cache is cleared, and unlocked.  On failure, performSelector: with a delay of half a second is set up to try again.  This is to prevent any kind of potential deadlock that might come up if the notification is received in the middle of RegexKitLite method, and that RKL method is what triggered the notification in the first place.  I think this is highly unlikely, but better safe than sorry.

C Pre-Processor Option 'RKL_PREPEND_TO_METHODS'

When set to something, ie -DRKL_PREPEND_TO_METHODS="xyz_", this causes all RegexKitLite methods to have whatever RKL_PREPEND_TO_METHODS is defined to to be added to the method name.  For exmaple, 'clearStringCache' would become 'xyz_clearStringCache'.  If RKL_PREPEND_TO_METHODS is not defined, nothing is prepended.  This is mainly for those users who prefer that user functionality added to the main Foundation classes, such as NSString, have a private prefix prepended to the function names to avoid conflicting with any system defined methods.



git-svn-id: https://regexkit.svn.sourceforge.net/svnroot/regexkit/RegexKitLite@32 83f0d1e4-963b-0410-a2a1-d3bd3f33e299