Most of the "changes" in this commit are due to a refactoring / renaming of some of the 4-way set associative cache.  This hit a lot of lines and probably represents 95% of the "changes", but it's all cosmetic.
There are a couple of very minor code refactoring changes that don't (or shouldn't) have any impact on how or what gets executed.  Things like tidying up a few casts reported by -Wc++-compat, warnings about unused labels when assertions are turned off, etc.

There is an important bug fix for ^Blocks support in this commit, though.  Found a problem when doing multi-threaded unit-tests with -DRKL_STACK_LIMIT=16 -DRKL_FIXED_LENGTH=8, which tends to "force" malloc() like allocations for a lot of things.  The function rkl_performEnumerationUsingBlock() was originally a bit of a clone of rkl_performRegexOp(), and with it came the use of the globally shared rkl_scratchBuffers[] for temporary buffer usage.  Well, rkl_performEnumerationUsingBlock() needs to operate outside the main rkl_spinLock, so this was causing a bit of a, uh, problem when two threads began using and free()ing the exact same buffers which they though they owned exclusively.  Fixed by adding scratchBuffers[] to the internal RKLBlockEnumerationHelper object, which is "enumeration local" copies of everything.  rkl_performEnumerationUsingBlocks sets up a RKLFindAll with **pointers to the RKLBlockEnumerationHelper enumeration local scratchBuffers[], and the RKLBlockEnumerationHelper -dealloc/-finalize cleans up the scratchBuffers[] (if neccessary) when the object is reaped.

Also fixed a memory leak in rkl_performEnumerationUsingBlock() when doing a search and replace like op- forgot to add a an -autorelease, and it was being missed by the unit tests because the bulk of the unit tests were using @"" constant strings, which wouldn't show up as a leak.  Fixed the bug and added/duplicated unit tests that use a NSMutableString, which would definitly tickle a leak.


git-svn-id: https://regexkit.svn.sourceforge.net/svnroot/regexkit/RegexKitLite@64 83f0d1e4-963b-0410-a2a1-d3bd3f33e299